Publish服务

+++ dao/model.go
@@ -3,14 +3,14 @@ package dao
 import "time"
 
 type User struct {
-	Id       int    `gorm:"primaryKey;autoincrement"`
+	Id       int64  `gorm:"primaryKey;autoincrement"`
 	Username string `gorm:"size:32;uniqueIndex;notnull"`
 	Password string `gorm:"size:32;notnull"`
 }
 
 type Video struct {
-	Id         int    `gorm:"primaryKey;autoincrement"`
-	AuthorId   int    `gorm:"notnull"`
+	Id         int64  `gorm:"primaryKey;autoincrement"`
+	AuthorId   int64  `gorm:"notnull"`
 	UploadTime int64  `gorm:"notnull;index"`
 	PlayUrl    string `gorm:"notnull;size:128" `
 	CoverUrl   string `gorm:"notnull;size:128" `
@@ -18,23 +18,23 @@ type Video struct {
 }
 
 type Favorite struct {
-	Id      int `gorm:"primaryKey;autoincrement"`
-	VideoId int `gorm:"notnull;unique index:idx_vid_uid"`
-	UserId  int `gorm:"notnull;unique index:idx_vid_uid;index"`
-	Status  int `gorm:"notnull;type:enum('1','2')"` // 1-点赞 2-未点赞
+	Id      int64 `gorm:"primaryKey;autoincrement"`
+	VideoId int64 `gorm:"notnull;unique index:idx_vid_uid"`
+	UserId  int64 `gorm:"notnull;unique index:idx_vid_uid;index"`
+	Status  int32 `gorm:"notnull;type:enum('1','2')"` // 1-点赞 2-未点赞
 }
 
 type Comment struct {
-	Id         int       `gorm:"primaryKey;autoincrement"`
-	VideoId    int       `gorm:"notnull;index"`
-	UserId     int       `gorm:"notnull;"`
+	Id         int64     `gorm:"primaryKey;autoincrement"`
+	VideoId    int64     `gorm:"notnull;index"`
+	UserId     int64     `gorm:"notnull;"`
 	CreateTime time.Time `gorm:"notnull"`
 	Content    string    `gorm:"notnull"`
 }
 
 type Relation struct {
-	Id         int `gorm:"primaryKey;autoincrement"`
-	FollowedId int `gorm:"notnull;unique index:idx_vid_uid"`       // 被关注
-	FollowerId int `gorm:"notnull;unique index:idx_vid_uid;index"` // 关注者
-	Status     int `gorm:"notnull;type:enum('1','2')"`             // 1-关注 2-未关注
+	Id         int64 `gorm:"primaryKey;autoincrement"`
+	FollowedId int64 `gorm:"notnull;unique index:idx_vid_uid"`       // 被关注
+	FollowerId int64 `gorm:"notnull;unique index:idx_vid_uid;index"` // 关注者
+	Status     int32 `gorm:"notnull;type:enum('1','2')"`             // 1-关注 2-未关注
 }
